in theory this fixes all SA1649 issues now with any project with stylecop enabled
