+++ src/modules/launcher/Wox.Infrastructure/Storage/IRepository`1.cs
@@ -1,0 +1,21 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Generic;
+
+namespace Wox.Infrastructure.Storage
+{
+    public interface IRepository<T>
+    {
+        void Add(T insertedItem);
+
+        void Remove(T removedItem);
+
+        bool Contains(T item);
+
+        void Set(IList<T> list);
+
+        bool Any();
+    }
+}
