+++ src/modules/launcher/Wox.Infrastructure/Storage/WoxJsonStorage`1.cs
@@ -6,7 +6,6 @@ using System.IO;
 
 namespace Wox.Infrastructure.Storage
 {
-    [System.Diagnostics.CodeAnalysis.SuppressMessage("StyleCop.CSharp.DocumentationRules", "SA1649:File name should match first type name", Justification = "Generic, file is named correctly")]
     public class WoxJsonStorage<T> : JsonStorage<T> where T : new()
     {
         public WoxJsonStorage()
