+++ src/modules/launcher/Wox.Infrastructure/UserSettings/CustomPluginHotkey.cs
